+++ applications/Unity.GrantManager/src/Unity.GrantManager.Application/Applicants/IApplicantService.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+using Unity.GrantManager.Applications;
+using Unity.GrantManager.Intakes;
+using Volo.Abp.Application.Services;
+
+namespace Unity.GrantManager.Applicants;
+
+public interface IApplicantsService : IApplicationService
+{
+    Task<Applicant> CreateOrRetrieveApplicantAsync(IntakeMapping intakeMap);
+    Task<ApplicantAgent> CreateOrUpdteApplicantAgentAsync(ApplicantAgentDto applicantAgentDto);
+}
